pyproject.toml
[tool.poetry]
name = "langx"
version = "0.1.0"
description = ""
authors = ["BuenVia <matthewclifford@hotmail.co.uk>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.4"
djangorestframework = "^3.15.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
